+++ TeenTam/account/models.py
@@ -11,7 +11,7 @@ class User(AbstractUser):
     birth = models.DateField(default=timezone.now)
     # 추가 기입 가능 필드
     profile_image = models.ImageField(
-        upload_to="profile_images", default="profile_images/no_profile_image.png")
+        upload_to="profile_images/%Y/%m/%d", default="profile_images/no_profile_image.png")
     postcode = models.IntegerField(null=True)
     address = models.CharField(max_length=200, null=True)
     detail_address = models.CharField(max_length=200, null=True)
